requirements.txt
uagents
requests
python-dotenv
openai

